Ensure we don't try to do things to an event we don't have
In theory we should get here because we should have stopped monitoring
the process, however in practice we do seem to be able to get to a
`:DOWN` callback and not having an event for this pid (which is probably
because it was already sent). So I guess in this case we just let it
slide and do nothing, since crashing here would lead to ruining all
events.