Fix a bug where during request start the pid might be gone away
and doing a Process.info(pid, :memory) returns a nil as a result.

This could happen if we have high load and handling e.g. a static
request, then by the time we get to ask about the pid that was dealing
with it, the process that was handling the request is already gone.

Perhaps, therefore as a result I should reconsider my stance on
calculating the durations myself and use the ones provided by the
measurements - since those are the ones that happened surrounding the
process and are not the durations between the times when we were
handling the start/end events