Replace :inets with Req
Basically it turns out that if events are large enough, doing the post
request with a super-basic-out-of-the-box-just-use-the-defaults :inets
no longer works, and one has to figure out the magical socket parameter
incantations to make it happen (probably this is when the data doesn't
really fit in a single packet or something).

While I would love to do that, it's ... work and honestly I don't feel
like going down the rabbit hole of tcp connection implementation details
so I will instead bit the bullet and have a dependency on an actual http
client.

This of course means that so will the users of this, but I suppose it's
better having this dep and def being able to send events, than spending
a bunch of time figuring out the correct socket options and then running
into more funky issues because I misunderstood something.